+++ src/mruby_integration/models/render_texture.cpp
@@ -12,36 +12,58 @@
 
 struct RClass* RenderTexture_class;
 
-void
-setup_RenderTexture(mrb_state* mrb, mrb_value object, int width, int height)
-{
-  mrb_iv_set(
-    mrb, object, mrb_intern_cstr(mrb, "@width"), mrb_int_value(mrb, width));
-  mrb_iv_set(
-    mrb, object, mrb_intern_cstr(mrb, "@height"), mrb_int_value(mrb, height));
-}
-
 auto
 mrb_RenderTexture_initialize(mrb_state* mrb, mrb_value self) -> mrb_value
 {
-  mrb_int width, height;
-  mrb_get_args(mrb, "ii", &width, &height);
+  // RenderTexture.new(width:, height:)
+  const mrb_int kw_num = 2;
+  const mrb_int kw_required = 2;
+  const mrb_sym kw_names[] = {
+    mrb_intern_lit(mrb, "width"),
+    mrb_intern_lit(mrb, "height"),
+  };
+  mrb_value kw_values[kw_num];
+  mrb_kwargs kwargs = { kw_num, kw_required, kw_names, kw_values, nullptr };
+  mrb_get_args(mrb, ":", &kwargs);
+
+  int width = 0;
+  if (!mrb_undef_p(kw_values[0])) {
+    width = mrb_as_int(mrb, kw_values[0]);
+  }
+
+  int height = 0;
+  if (!mrb_undef_p(kw_values[1])) {
+    height = mrb_as_int(mrb, kw_values[1]);
+  }
 
   auto* render_texture =
     static_cast<RenderTexture*>(malloc(sizeof(RenderTexture)));
   *render_texture = LoadRenderTexture(width, height);
 
-  setup_RenderTexture(mrb, self, width, height);
-
-  Texture2D* texture = &render_texture->texture;
-
-  ivar_attr_texture2d(mrb, self, render_texture->texture, texture);
-
   add_reference(&render_texture->texture);
+  mrb_iv_set(mrb,
+             self,
+             mrb_intern_cstr(mrb, "@texture"),
+             mrb_Texture2D_value(mrb, &render_texture->texture));
 
   mrb_data_init(self, render_texture, &RenderTexture_type);
   return self;
 }
+
+mrb_attr_reader_with_klasses_and_attrs(mrb,
+                                       self,
+                                       int,
+                                       RenderTexture,
+                                       RenderTexture,
+                                       width,
+                                       texture.width);
+mrb_attr_reader_with_klasses_and_attrs(mrb,
+                                       self,
+                                       int,
+                                       RenderTexture,
+                                       RenderTexture,
+                                       height,
+                                       texture.height);
 
 auto
 mrb_RenderTexture_unload(mrb_state* mrb, mrb_value self) -> mrb_value
@@ -66,6 +88,8 @@ append_models_RenderTexture(mrb_state* mrb)
                     "initialize",
                     mrb_RenderTexture_initialize,
                     MRB_ARGS_REQ(2));
+  mrb_attr_reader_define(mrb, RenderTexture, width);
+  mrb_attr_reader_define(mrb, RenderTexture, height);
   mrb_define_method(mrb,
                     RenderTexture_class,
                     "unload",